Reorganizing tests and improving tests performance
